création des managers

+++ src/org/eni/encheres/bo/Article.java
@@ -8,12 +8,12 @@ public class Article {
 	private String description;
 	private LocalDateTime dateDebutEncheres;
 	private LocalDateTime dateFinEncheres;
-	private float miseAPrix;
-	private float prixVente;
+	private int miseAPrix;
+	private int prixVente;
 	private String etatVente;
 	
 	public Article(int noArticle, String nomArticle, String description, LocalDateTime dateDebutEncheres,
-			LocalDateTime dateFinEncheres, float miseAPrix, float prixVente, String etatVente) {
+			LocalDateTime dateFinEncheres, int miseAPrix, int prixVente, String etatVente) {
 		super();
 		this.noArticle = noArticle;
 		this.nomArticle = nomArticle;
@@ -25,7 +25,7 @@ public class Article {
 		this.etatVente = etatVente;
 	}
 	public Article(String nomArticle, String description, LocalDateTime dateDebutEncheres,
-			LocalDateTime dateFinEncheres, float miseAPrix) {
+			LocalDateTime dateFinEncheres, int miseAPrix) {
 		super();
 		this.nomArticle = nomArticle;
 		this.description = description;
@@ -63,16 +63,16 @@ public class Article {
 	public void setDateFinEncheres(LocalDateTime dateFinEncheres) {
 		this.dateFinEncheres = dateFinEncheres;
 	}
-	public float getMiseAPrix() {
+	public int getMiseAPrix() {
 		return miseAPrix;
 	}
-	public void setMiseAPrix(float miseAPrix) {
+	public void setMiseAPrix(int miseAPrix) {
 		this.miseAPrix = miseAPrix;
 	}
-	public float getPrixVente() {
+	public int getPrixVente() {
 		return prixVente;
 	}
-	public void setPrixVente(float prixVente) {
+	public void setPrixVente(int prixVente) {
 		this.prixVente = prixVente;
 	}
 	public String getEtatVente() {
